Added a receiver to monitor network connection changes
The DataUsageService should only run when connected to a Cellular
network and not on WiFi. The DataUsageService is only scheduled to
be run if connected to a metered network. A network state change
receiver monitors networks state changes and starts the scheduling
alarm if the new network is metered and the DataUsageService is
enabled.

Change-Id: If01385805c4e61aea7f0b7788de84a65a8b6c08b